requirements.txt
pytest
selenium
jsonpickle
pymysql
lxml-4.5.2-cp38-cp38-win32.whl
allure-pytest